move process_flat to rfuns
